Day 10 Part 2 now runs in 17ms under Release (~2s in Debug)
I tried different optimizations of the old approach even trying direct
recursion to avoid the RAM explosion, but that run never finished.

I gave up and implemented an equation solver and ... got stuck for a
full day, because I had a bug in my counting logic, which resulted in
the solver not trying out all variable values and it failed only for the most
complex inputs (the ones with 13 buttons!). Took me way to long to
figure that one out >.<

Added my testing data for the input that caused me most of the headache.